Rename model and view to uppercase
