+++ js/SliderModel.js
