+++ js/SliderView.js
