+++ js/adapt-contrib-slider.js
@@ -1,6 +1,6 @@
 import components from 'core/js/components';
-import SliderView from './sliderView';
-import SliderModel from './sliderModel';
+import SliderView from './SliderView';
+import SliderModel from './SliderModel';
 
 export default components.register('slider', {
   view: SliderView,
